requirements.txt
streamlit
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
pydantic
google-genai
python-dotenv
google-api-python-client
